+++ layers/GatoGraphQLForWP/plugins/gatographql/src/ModuleResolvers/Extensions/PowerBundleExtensionModuleResolver.php
@@ -11,11 +11,11 @@ use GatoGraphQL\GatoGraphQL\Plugin;
 use GatoGraphQL\GatoGraphQL\PluginApp;
 use GatoGraphQL\GatoGraphQL\PluginStaticModuleConfiguration;
 
-class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolver
+class PowerBundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolver
 {
-    public const PRO = Plugin::NAMESPACE . '\\bundle-extensions\\pro';
+    // public const PRO = Plugin::NAMESPACE . '\\bundle-extensions\\pro';
 
-    public const ALL_EXTENSIONS = Plugin::NAMESPACE . '\\bundle-extensions\\all-extensions';
+    // public const ALL_EXTENSIONS = Plugin::NAMESPACE . '\\bundle-extensions\\all-extensions';
     public const POWER_EXTENSIONS = Plugin::NAMESPACE . '\\bundle-extensions\\power-extensions';
 
     public const ACCESS_CONTROL = Plugin::NAMESPACE . '\\bundle-extensions\\access-control';
@@ -26,7 +26,6 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     public const INTERNAL_GRAPHQL_SERVER = Plugin::NAMESPACE . '\\bundle-extensions\\internal-graphql-server';
     public const MULTIPLE_QUERY_EXECUTION = Plugin::NAMESPACE . '\\bundle-extensions\\multiple-query-execution';
     public const PERSISTED_QUERIES = Plugin::NAMESPACE . '\\bundle-extensions\\persisted-queries';
-    public const POLYLANG_INTEGRATION = Plugin::NAMESPACE . '\\bundle-extensions\\polylang-integration';
     public const QUERY_FUNCTIONS = Plugin::NAMESPACE . '\\bundle-extensions\\query-functions';
     public const SCHEMA_FUNCTIONS = Plugin::NAMESPACE . '\\bundle-extensions\\schema-functions';
 
@@ -36,9 +35,9 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     public function getModulesToResolve(): array
     {
         return array_merge(
-            PluginStaticModuleConfiguration::displayGatoGraphQLPROBundleOnExtensionsPage() ? [
-                self::PRO,
-            ] : [],
+            // PluginStaticModuleConfiguration::displayGatoGraphQLPROBundleOnExtensionsPage() ? [
+            //     self::PRO,
+            // ] : [],
             PluginStaticModuleConfiguration::displayGatoGraphQLPROFeatureBundlesOnExtensionsPage() ? [
                 self::ACCESS_CONTROL,
                 self::CACHING,
@@ -48,12 +47,11 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
                 self::INTERNAL_GRAPHQL_SERVER,
                 self::MULTIPLE_QUERY_EXECUTION,
                 self::PERSISTED_QUERIES,
-                self::POLYLANG_INTEGRATION,
                 self::QUERY_FUNCTIONS,
                 self::SCHEMA_FUNCTIONS,
             ] : [],
             PluginStaticModuleConfiguration::displayGatoGraphQLPROAllExtensionsBundleOnExtensionsPage() ? [
-                // self::ALL_EXTENSIONS,
+                // // self::ALL_EXTENSIONS,
                 self::POWER_EXTENSIONS,
             ] : [],
         );
@@ -64,10 +62,10 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
         $bundlePlaceholder = \__('“%s” bundle', 'gatographql');
         $extensionPlaceholder = \__('%s', 'gatographql');
         return match ($module) {
-            self::PRO => \__('Gato GraphQL PRO', 'gatographql'),
-            self::ALL_EXTENSIONS => sprintf($bundlePlaceholder, \__('All Extensions', 'gatographql')),
+            // self::PRO => \__('Gato GraphQL PRO', 'gatographql'),
+            // self::ALL_EXTENSIONS => sprintf($bundlePlaceholder, \__('All Extensions', 'gatographql')),
             self::POWER_EXTENSIONS => sprintf($bundlePlaceholder, \__('Power Extensions', 'gatographql')),
-            
+
             self::ACCESS_CONTROL => sprintf($extensionPlaceholder, \__('Access Control', 'gatographql')),
             self::CACHING => sprintf($extensionPlaceholder, \__('Caching', 'gatographql')),
             self::CUSTOM_ENDPOINTS => sprintf($extensionPlaceholder, \__('Custom Endpoints', 'gatographql')),
@@ -76,7 +74,6 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
             self::INTERNAL_GRAPHQL_SERVER => sprintf($extensionPlaceholder, \__('Internal GraphQL Server', 'gatographql')),
             self::MULTIPLE_QUERY_EXECUTION => sprintf($extensionPlaceholder, \__('Multiple Query Execution', 'gatographql')),
             self::PERSISTED_QUERIES => sprintf($extensionPlaceholder, \__('Persisted Queries', 'gatographql')),
-            self::POLYLANG_INTEGRATION => sprintf($extensionPlaceholder, \__('Polylang Integration', 'gatographql')),
             self::QUERY_FUNCTIONS => sprintf($extensionPlaceholder, \__('Query Functions', 'gatographql')),
             self::SCHEMA_FUNCTIONS => sprintf($extensionPlaceholder, \__('Schema Functions', 'gatographql')),
             default => $module,
@@ -86,10 +83,10 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     public function getDescription(string $module): string
     {
         return match ($module) {
-            self::PRO => \__('All the PRO extensions for Gato GraphQL, the most powerful GraphQL server for WordPress', 'gatographql'),
-            self::ALL_EXTENSIONS => \__('All of Gato GraphQL extensions, in a single plugin', 'gatographql'),
+            // self::PRO => \__('All the PRO extensions for Gato GraphQL, the most powerful GraphQL server for WordPress', 'gatographql'),
+            // self::ALL_EXTENSIONS => \__('All of Gato GraphQL extensions, in a single plugin', 'gatographql'),
             self::POWER_EXTENSIONS => \__('All of Gato GraphQL\'s power extensions, in a single plugin', 'gatographql'),
-            
+
             self::ACCESS_CONTROL => \__('Define Access Control Lists to manage granular access to the API for your users', 'gatographql'),
             self::CACHING => \__('Make your application faster by providing HTTP Caching for the GraphQL response, and by caching the results of expensive operations', 'gatographql'),
             self::CUSTOM_ENDPOINTS => \__('Create custom schemas, with custom access rules for different users, each available under its own endpoint', 'gatographql'),
@@ -98,7 +95,6 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
             self::INTERNAL_GRAPHQL_SERVER => \__('Execute GraphQL queries directly within your application, using PHP code', 'gatographql'),
             self::MULTIPLE_QUERY_EXECUTION => \__('Combine multiple GraphQL queries together, and execute them as a single operation, to improve performance and make your queries more manageable', 'gatographql'),
             self::PERSISTED_QUERIES => \__('Use GraphQL queries to create pre-defined endpoints as in REST, obtaining the benefits from both APIs', 'gatographql'),
-            self::POLYLANG_INTEGRATION => \__('Integration with the Polylang plugin, providing fields to the GraphQL schema to fetch multilingual data', 'gatographql'),
             self::QUERY_FUNCTIONS => \__('Manipulate the values of fields within the GraphQL query, via a collection of utilities and special directives providing meta-programming capabilities', 'gatographql'),
             self::SCHEMA_FUNCTIONS => \__('Collection of fields and directives added to the GraphQL schema, providing useful functionality', 'gatographql'),
             default => parent::getDescription($module),
@@ -107,7 +103,7 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
 
     public function getPriority(): int
     {
-        return 20;
+        return 30;
     }
 
     public function getWebsiteURL(string $module): string
@@ -115,8 +111,8 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
         /** @var ModuleConfiguration */
         $moduleConfiguration = App::getModule(Module::class)->getConfiguration();
         return match ($module) {
-            self::PRO,
-            self::ALL_EXTENSIONS,
+            // self::PRO,
+            // self::ALL_EXTENSIONS,
             self::POWER_EXTENSIONS =>
                 $moduleConfiguration->getGatoGraphQLWebsiteURL(),
             default
@@ -128,8 +124,8 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     {
         $pluginURL = PluginApp::getMainPlugin()->getPluginURL();
         return match ($module) {
-            self::PRO,
-            self::ALL_EXTENSIONS,
+            // self::PRO,
+            // self::ALL_EXTENSIONS,
             self::POWER_EXTENSIONS
                 => $pluginURL . 'assets/img/logos/GatoGraphQL-logo-face.png',
             self::ACCESS_CONTROL
@@ -148,8 +144,6 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
                 => $pluginURL . 'assets/img/extension-logos/multiple-query-execution.svg',
             self::PERSISTED_QUERIES
                 => $pluginURL . 'assets/img/extension-logos/persisted-queries.svg',
-            self::POLYLANG_INTEGRATION
-                => $pluginURL . 'assets/img/extension-logos/polylang-integration.png',
             self::QUERY_FUNCTIONS
                 => $pluginURL . 'assets/img/extension-logos/query-functions.svg',
             self::SCHEMA_FUNCTIONS
@@ -164,68 +158,64 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     public function getBundledExtensionModules(string $module): array
     {
         return match ($module) {
-            self::PRO => [
-                ExtensionModuleResolver::ACCESS_CONTROL,
-                ExtensionModuleResolver::ACCESS_CONTROL_VISITOR_IP,
-                ExtensionModuleResolver::AUTOMATION,
-                ExtensionModuleResolver::CACHE_CONTROL,
-                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
-                ExtensionModuleResolver::CUSTOM_ENDPOINTS,
-                ExtensionModuleResolver::DEPRECATION_NOTIFIER,
-                ExtensionModuleResolver::EMAIL_SENDER,
-                ExtensionModuleResolver::EVENTS_MANAGER,
-                ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
-                ExtensionModuleResolver::FIELD_DEPRECATION,
-                ExtensionModuleResolver::FIELD_ON_FIELD,
-                ExtensionModuleResolver::FIELD_RESOLUTION_CACHING,
-                ExtensionModuleResolver::FIELD_RESPONSE_REMOVAL,
-                ExtensionModuleResolver::FIELD_TO_INPUT,
-                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
-                ExtensionModuleResolver::GOOGLE_TRANSLATE,
-                ExtensionModuleResolver::HELPER_FUNCTION_COLLECTION,
-                ExtensionModuleResolver::HTTP_CLIENT,
-                ExtensionModuleResolver::HTTP_REQUEST_VIA_SCHEMA,
-                ExtensionModuleResolver::INTERNAL_GRAPHQL_SERVER,
-                ExtensionModuleResolver::LOW_LEVEL_PERSISTED_QUERY_EDITING,
-                ExtensionModuleResolver::MULTILINGUALPRESS,
-                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
-                ExtensionModuleResolver::PERSISTED_QUERIES,
-                ExtensionModuleResolver::PHP_CONSTANTS_AND_ENVIRONMENT_VARIABLES_VIA_SCHEMA,
-                ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
-                ExtensionModuleResolver::POLYLANG,
-                ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
-            ],
-            self::ALL_EXTENSIONS => [
-                ExtensionModuleResolver::ACCESS_CONTROL,
-                ExtensionModuleResolver::ACCESS_CONTROL_VISITOR_IP,
-                // ExtensionModuleResolver::AUTOMATION,
-                ExtensionModuleResolver::CACHE_CONTROL,
-                ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
-                ExtensionModuleResolver::CUSTOM_ENDPOINTS,
-                ExtensionModuleResolver::DEPRECATION_NOTIFIER,
-                ExtensionModuleResolver::EMAIL_SENDER,
-                // ExtensionModuleResolver::EVENTS_MANAGER,
-                ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
-                ExtensionModuleResolver::FIELD_DEPRECATION,
-                ExtensionModuleResolver::FIELD_ON_FIELD,
-                ExtensionModuleResolver::FIELD_RESOLUTION_CACHING,
-                ExtensionModuleResolver::FIELD_RESPONSE_REMOVAL,
-                ExtensionModuleResolver::FIELD_TO_INPUT,
-                ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
-                // ExtensionModuleResolver::GOOGLE_TRANSLATE,
-                ExtensionModuleResolver::HELPER_FUNCTION_COLLECTION,
-                ExtensionModuleResolver::HTTP_CLIENT,
-                ExtensionModuleResolver::HTTP_REQUEST_VIA_SCHEMA,
-                ExtensionModuleResolver::INTERNAL_GRAPHQL_SERVER,
-                ExtensionModuleResolver::LOW_LEVEL_PERSISTED_QUERY_EDITING,
-                // ExtensionModuleResolver::MULTILINGUALPRESS,
-                ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
-                ExtensionModuleResolver::PERSISTED_QUERIES,
-                ExtensionModuleResolver::PHP_CONSTANTS_AND_ENVIRONMENT_VARIABLES_VIA_SCHEMA,
-                ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
-                ExtensionModuleResolver::POLYLANG,
-                ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
-            ],
+            // self::PRO => [
+            //     ExtensionModuleResolver::ACCESS_CONTROL,
+            //     ExtensionModuleResolver::ACCESS_CONTROL_VISITOR_IP,
+            //     ExtensionModuleResolver::AUTOMATION,
+            //     ExtensionModuleResolver::CACHE_CONTROL,
+            //     ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+            //     ExtensionModuleResolver::CUSTOM_ENDPOINTS,
+            //     ExtensionModuleResolver::DEPRECATION_NOTIFIER,
+            //     ExtensionModuleResolver::EMAIL_SENDER,
+            //     ExtensionModuleResolver::EVENTS_MANAGER,
+            //     ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
+            //     ExtensionModuleResolver::FIELD_DEPRECATION,
+            //     ExtensionModuleResolver::FIELD_ON_FIELD,
+            //     ExtensionModuleResolver::FIELD_RESOLUTION_CACHING,
+            //     ExtensionModuleResolver::FIELD_RESPONSE_REMOVAL,
+            //     ExtensionModuleResolver::FIELD_TO_INPUT,
+            //     ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
+            //     ExtensionModuleResolver::GOOGLE_TRANSLATE,
+            //     ExtensionModuleResolver::HELPER_FUNCTION_COLLECTION,
+            //     ExtensionModuleResolver::HTTP_CLIENT,
+            //     ExtensionModuleResolver::HTTP_REQUEST_VIA_SCHEMA,
+            //     ExtensionModuleResolver::INTERNAL_GRAPHQL_SERVER,
+            //     ExtensionModuleResolver::LOW_LEVEL_PERSISTED_QUERY_EDITING,
+            //     ExtensionModuleResolver::MULTILINGUALPRESS,
+            //     ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+            //     ExtensionModuleResolver::PERSISTED_QUERIES,
+            //     ExtensionModuleResolver::PHP_CONSTANTS_AND_ENVIRONMENT_VARIABLES_VIA_SCHEMA,
+            //     ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
+            //     ExtensionModuleResolver::POLYLANG,
+            //     ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
+            // ],
+            // self::ALL_EXTENSIONS => [
+            //     ExtensionModuleResolver::ACCESS_CONTROL,
+            //     ExtensionModuleResolver::ACCESS_CONTROL_VISITOR_IP,
+            //     ExtensionModuleResolver::CACHE_CONTROL,
+            //     ExtensionModuleResolver::CONDITIONAL_FIELD_MANIPULATION,
+            //     ExtensionModuleResolver::CUSTOM_ENDPOINTS,
+            //     ExtensionModuleResolver::DEPRECATION_NOTIFIER,
+            //     ExtensionModuleResolver::EMAIL_SENDER,
+            //     ExtensionModuleResolver::FIELD_DEFAULT_VALUE,
+            //     ExtensionModuleResolver::FIELD_DEPRECATION,
+            //     ExtensionModuleResolver::FIELD_ON_FIELD,
+            //     ExtensionModuleResolver::FIELD_RESOLUTION_CACHING,
+            //     ExtensionModuleResolver::FIELD_RESPONSE_REMOVAL,
+            //     ExtensionModuleResolver::FIELD_TO_INPUT,
+            //     ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
+            //     ExtensionModuleResolver::HELPER_FUNCTION_COLLECTION,
+            //     ExtensionModuleResolver::HTTP_CLIENT,
+            //     ExtensionModuleResolver::HTTP_REQUEST_VIA_SCHEMA,
+            //     ExtensionModuleResolver::INTERNAL_GRAPHQL_SERVER,
+            //     ExtensionModuleResolver::LOW_LEVEL_PERSISTED_QUERY_EDITING,
+            //     ExtensionModuleResolver::MULTIPLE_QUERY_EXECUTION,
+            //     ExtensionModuleResolver::PERSISTED_QUERIES,
+            //     ExtensionModuleResolver::PHP_CONSTANTS_AND_ENVIRONMENT_VARIABLES_VIA_SCHEMA,
+            //     ExtensionModuleResolver::PHP_FUNCTIONS_VIA_SCHEMA,
+            //     ExtensionModuleResolver::POLYLANG,
+            //     ExtensionModuleResolver::RESPONSE_ERROR_TRIGGER,
+            // ],
             self::POWER_EXTENSIONS => [
                 ExtensionModuleResolver::ACCESS_CONTROL,
                 ExtensionModuleResolver::ACCESS_CONTROL_VISITOR_IP,
@@ -280,9 +270,6 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
                 ExtensionModuleResolver::PERSISTED_QUERIES,
                 ExtensionModuleResolver::LOW_LEVEL_PERSISTED_QUERY_EDITING,
             ],
-            self::POLYLANG_INTEGRATION => [
-                ExtensionModuleResolver::POLYLANG,
-            ],
             self::QUERY_FUNCTIONS => [
                 ExtensionModuleResolver::FIELD_TO_INPUT,
                 ExtensionModuleResolver::FIELD_VALUE_ITERATION_AND_MANIPULATION,
@@ -309,24 +296,24 @@ class BundleExtensionModuleResolver extends AbstractBundleExtensionModuleResolve
     public function getBundledBundleExtensionModules(string $module): array
     {
         return match ($module) {
-            // "Gato GraphQL PRO" bundles all other bundles
-            self::PRO => array_diff(
-                $this->getModulesToResolve(),
-                [$module]
-            ),
-            self::ALL_EXTENSIONS => [
-                self::ACCESS_CONTROL,
-                self::CACHING,
-                self::CUSTOM_ENDPOINTS,
-                self::DEPRECATION,
-                self::HTTP_CLIENT,
-                self::INTERNAL_GRAPHQL_SERVER,
-                self::MULTIPLE_QUERY_EXECUTION,
-                self::PERSISTED_QUERIES,
-                self::POLYLANG_INTEGRATION,
-                self::QUERY_FUNCTIONS,
-                self::SCHEMA_FUNCTIONS,
-            ],
+            // // "Gato GraphQL PRO" bundles all other bundles
+            // self::PRO => array_diff(
+            //     $this->getModulesToResolve(),
+            //     [$module]
+            // ),
+            // self::ALL_EXTENSIONS => [
+            //     self::ACCESS_CONTROL,
+            //     self::CACHING,
+            //     self::CUSTOM_ENDPOINTS,
+            //     self::DEPRECATION,
+            //     self::HTTP_CLIENT,
+            //     self::INTERNAL_GRAPHQL_SERVER,
+            //     self::MULTIPLE_QUERY_EXECUTION,
+            //     self::PERSISTED_QUERIES,
+            //     PremiumBundleExtensionModuleResolver::POLYLANG_INTEGRATION,
+            //     self::QUERY_FUNCTIONS,
+            //     self::SCHEMA_FUNCTIONS,
+            // ],
             self::POWER_EXTENSIONS => [
                 self::ACCESS_CONTROL,
                 self::CACHING,