+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -14,7 +14,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 - GraphQL type `StringListValueJSONObject` (#2991)
 - YouTube videos to the extension docs (#9b95be5)
-- Documentation for new DeepL extension
+- Documentation for Premium Extensions (#2994)
 
 ### Improvements